+++ src/models/Mediafile.php
@@ -184,20 +184,6 @@ class Mediafile extends ActiveRecord
     }
 
     /**
-     * Remove this mediafile owner.
-     *
-     * @param int    $ownerId
-     * @param string $owner
-     * @param string $ownerAttribute
-     *
-     * @return bool
-     */
-    public static function removeOwner(int $ownerId, string $owner, string $ownerAttribute): bool
-    {
-        return OwnerMediafile::removeOwner($ownerId, $owner, $ownerAttribute);
-    }
-
-    /**
      * @return \yii\db\ActiveQuery
      */
     public function getOwners()